requirements/local.txt
# Local dev dependencies, not to be deployed
-r base.txt
coverage==3.7.1
django-debug-toolbar==1.0.1
django-dotenv==1.2
Sphinx==1.2.1
